+++ agent-langgraph-stateful-short-term/.env.example
@@ -9,12 +9,10 @@ DATABRICKS_CONFIG_PROFILE=DEFAULT
 # TODO: Update with the MLflow experiment you want to log traces and models to
 MLFLOW_EXPERIMENT_ID=
 
-# TODO: Update with your Lakebase instance name for short-term memory / conversation checkpointing
-LAKEBASE_INSTANCE_NAME=
-
 # TODO: Update the route to query agent if you used a different port to deploy your agent
 API_PROXY=http://localhost:8000/invocations
 
 CHAT_APP_PORT=3000
+CHAT_PROXY_TIMEOUT_SECONDS=300
 MLFLOW_TRACKING_URI="databricks"
 MLFLOW_REGISTRY_URI="databricks-uc"
