+++ agent-langgraph-stateful-short-term/pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "agent-server"
 version = "0.1.0"
-description = "MLflow-compatible agent server with short-term memory via Lakebase"
+description = "MLflow-compatible agent server with FastAPI"
 readme = "README.md"
 authors = [
     { name = "Agent Developer", email = "developer@example.com" }
@@ -10,11 +10,15 @@ requires-python = ">=3.11"
 dependencies = [
     "fastapi>=0.115.12",
     "uvicorn>=0.34.2",
-    "databricks-langchain[memory]>=0.11.0",
-    "mlflow>=3.6.0",
+    "databricks-langchain[memory]>=0.12.0",
+    "mlflow>=3.8.0rc0",
     "langgraph>=1.0.1",
+    "langchain-mcp-adapters>=0.1.11",
     "python-dotenv",
 ]
+
+[tool.uv]
+prerelease = "allow"
 
 [build-system]
 requires = ["hatchling"]
@@ -30,4 +34,3 @@ dev = [
 [project.scripts]
 start-server = "agent_server.start_server:main"
 agent-evaluate = "agent_server.evaluate_agent:evaluate"
-
