+++ agent-langgraph-stateful-short-term/scripts/start-app.sh
@@ -28,7 +28,7 @@ if [ ! -d "e2e-chatbot-app-next" ]; then
         echo "ERROR: Failed to clone repository."
         echo "Please manually download the folder by going to the following link:"
         echo "  https://download-directory.github.io/?url=https://github.com/databricks/app-templates/tree/main/e2e-chatbot-app-next"
-        echo "Then unzip it in this directory and re-run \`./scripts/start-app.sh\`."
+        echo "Then unzip it in this directory and re-run `./scripts/start-app.sh`."
         exit 1
     fi
     
@@ -100,4 +100,3 @@ echo "=========================================="
 print_error_logs
 cleanup
 exit $EXIT_CODE
-
